requirements.txt
sentence-transformers
faiss-cpu
PyMuPDF
requests
langchain
llama-cpp-python
flask
